move sparqlToAST into generateVisualQuery
